Add origin var., re-do chi.sq, re-do prop. plots -
I loaded all libraries at the top of the script. Created a new variable "origin" to avoid having to use "Otolith Mark Present" in figures and tables, re-did the chi square tests to compare the proportion of parents by origin to the proportion of offspring by parental-origin. Re-did the proportion plots for family size so that each sex/origin adds up to one within a plot. The idea behind that is that we want to know what proportion of say hatchery males produced 0, 1, 2, etc. offspring. Updated script so that the whole thing runs from the top (no dependencies on objects created lower in the notebook). Also, as an aesthetic thing, I reformated the anova and lm calls so that we specify the data and then specify the variables rather than data$variable, this makes things easier to read.